added TinyMCE to stimulus fields and MC/MR feedback
At present when removing or adding an option to an MC/MR item the feedback
TinyMCEs are switched off, then the row is added or removed, then they're all
switched on again. This is very slow but it's the only solution I can find at
present. I think the problem is that otherwise various elements aren't re-IDd
properly.